+++ context/app/static/js/components/Search/Search.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import PortalSearch from '@hubmap/portal-search';
 import Container from '@material-ui/core/Container';
+import PortalSearch from '../PortalSearch/PortalSearch';
 import { readCookie } from '../../helpers/functions';
 import 'searchkit/theming/theme.scss';
 import './Search.css';
@@ -143,6 +143,7 @@ const searchProps = {
   // "type" should be one of the filters described here:
   // http://docs.searchkit.co/stable/components/navigation/
   filters: filtersByType[type],
+  hiddenFilterIds: ['entity_type'],
 };
 
 function Search(props) {
@@ -151,7 +152,7 @@ function Search(props) {
   /* eslint-disable react/jsx-props-no-spreading */
   return (
     <Container maxWidth="lg">
-      <PortalSearch.Search {...allProps} />
+      <PortalSearch {...allProps} />
     </Container>
   );
 /* eslint-enable react/jsx-props-no-spreading */